chore: delete lot of code
