--- xtask/bench/src/analyzer-libs-js.txt
@@ -1,2 +1,0 @@
-https://raw.githubusercontent.com/prettier/prettier/main/src/language-js/utils/index.js
-https://raw.githubusercontent.com/jquery/jquery/main/src/css.js
