--- xtask/bench/src/analyzer-libs-ts.txt
@@ -1,5 +1,0 @@
-https://raw.githubusercontent.com/angular/angular/master/packages/compiler/src/expression_parser/parser.ts
-https://raw.githubusercontent.com/angular/angular/master/packages/router/src/router.ts
-https://raw.githubusercontent.com/rome/tools/archived-js/internal/js-parser/parser/statement.ts
-https://raw.githubusercontent.com/rome/tools/archived-js/internal/diagnostics/descriptions/lint.ts
-https://raw.githubusercontent.com/rome/tools/archived-js/internal/js-parser/parser/typescript.ts
