--- xtask/bench/src/libs-css.txt
@@ -1,8 +1,0 @@
-https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.css
-https://cdn.jsdelivr.net/npm/daisyui@3.9.2/dist/full.css
-https://cdn.jsdelivr.net/npm/purecss@3.0.0/build/pure.css
-https://raw.githubusercontent.com/Dogfalo/materialize/v1-dev/dist/css/materialize.css
-https://raw.githubusercontent.com/jgthms/bulma/master/css/bulma.css
-https://raw.githubusercontent.com/foundation/foundation-sites/develop/dist/css/foundation.css
-https://raw.githubusercontent.com/Semantic-Org/Semantic-UI/master/dist/semantic.css
-https://raw.githubusercontent.com/tachyons-css/tachyons/main/css/tachyons.css
